Introduce CreditAccount in credit
