+++ 1 Hands-on Legacy Code/test/application/AccountManagementServiceTest.java
@@ -8,10 +8,11 @@ import java.time.LocalDate;
 import java.util.Iterator;
 import java.util.Set;
 
+import accounting.AccountManagementService;
 import org.junit.jupiter.api.Test;
 
 import models.Account;
-import models.Credit;
+import credit.Credit;
 import models.CreditAccount;
 import models.Customer;
